requirements_mcp.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.16
openai>=1.3.5
cohere>=5.5.8
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
mcp-tools>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0